Changed rudimentary level generation to room layout based on small rooms
Now, the level is generated by first generating small rooms next
to each other. They are so simple that many checks regarding
door placement are not necessary in the first place. Furthermore,
it is easier to enforce additional constraints on a sequence of
1x1 rooms. After the basic generation, multiple rooms are combined
if the pattern of a larger room form factor fits onto the grid and
if the cluster of rooms is subsequently selected to actually be
replaced by that very room.