Implemented multi-stage aggregation based level generation from RoomGrid
Initially, the RoomGrid is populated by small rooms covering one grid cell.
They are said to have one section. Each section points to another section
or small room.

In another stage, possible aggregations of small rooms to form bigger rooms
are searched for. They are selected, so that they don't overlap. Then, the
connectivity graph of the small rooms is condensed and doors are removed that
would point from a bigger room into itself. The remaining smaller rooms are
left as they are.